+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to
 
 - `Header` component
 
+### Fixed
+
+- Made `Container` use `<div>` instead of `<span>`
+
 ## [0.0.3] - 10-12-2019
 
 ### Added